Cleaned up the code and changed the read me
